requirements/base.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.10.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.12.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
zipp==3.8.1
celery==5.2.7
redis==4.3.4
